added area calc + remove constructor body

+++ src/main/java/com/kodilla/testing/shape/Shape.java
@@ -3,5 +3,5 @@ package com.kodilla.testing.shape;
 interface Shape {
 
     String getShapeName();
-    Integer getField();
+    Double getField();
 }
